requirements.txt
requests
bs4
flask
sqlalchemy
alembic
waitress

fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings